+++ src/lib.rs
@@ -1,39 +1,3 @@
-use actix_web::dev::Server;
-use actix_web::{web, App, HttpResponse, HttpServer};
-use std::net::TcpListener;
-
-#[derive(serde::Deserialize)]
-struct FormData {
-    email: String,
-    name: String,
-}
-
-async fn check_health() -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-async fn subscriptions(_form: web::Form<FormData>) -> HttpResponse {
-    HttpResponse::Ok().finish()
-}
-
-pub fn start(listener: TcpListener) -> Result<Server, std::io::Error> {
-    let server = HttpServer::new(|| {
-        App::new()
-            .route("/checkhealth", web::get().to(check_health))
-            .route("/subscriptions", web::post().to(subscriptions))
-    })
-    .listen(listener)?
-    .run();
-
-    Ok(server)
-}
-
-pub fn start_background() -> String {
-    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
-    let port = listener.local_addr().unwrap().port();
-
-    let server = start(listener).expect("Failed to bind address");
-    tokio::spawn(server);
-
-    format!("http://127.0.0.1:{}", port)
-}
+pub mod configuration;
+pub mod routes;
+pub mod startup;
