+++ src/routes/check_health.rs
@@ -1,0 +1,6 @@
+use actix_web::HttpResponse;
+
+
+pub async fn check_health() -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
