+++ src/routes/mod.rs
@@ -1,0 +1,5 @@
+mod check_health;
+mod subscriptions;
+
+pub use check_health::*;
+pub use subscriptions::*;
