+++ src/routes/subscriptions.rs
@@ -1,0 +1,12 @@
+use actix_web::{web, HttpResponse};
+
+
+#[derive(serde::Deserialize)]
+pub struct FormData {
+    email: String,
+    name: String,
+}
+
+pub async fn subscriptions(_form: web::Form<FormData>) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
