+++ src/startup.rs
@@ -1,0 +1,27 @@
+use std::net::TcpListener;
+use actix_web::dev::Server;
+use actix_web::{web, App, HttpServer};
+
+use crate::routes::*;
+
+pub fn start(listener: TcpListener) -> Result<Server, std::io::Error> {
+    let server = HttpServer::new(|| {
+        App::new()
+            .route("/check_health", web::get().to(check_health))
+            .route("/subscriptions", web::post().to(subscriptions))
+    })
+    .listen(listener)?
+    .run();
+
+    Ok(server)
+}
+
+pub fn start_background() -> String {
+    let listener = TcpListener::bind("127.0.0.1:0").unwrap();
+    let port = listener.local_addr().unwrap().port();
+
+    let server = start(listener).expect("Failed to bind address");
+    tokio::spawn(server);
+
+    format!("http://127.0.0.1:{}", port)
+}
